fix package export paths and add types
