+++ src/components/menubar/readme.md
@@ -10,11 +10,11 @@ When a main menubar item is the current active one, a sub-menubar is shown and e
 
 ## Properties
 
-| Property      | Attribute      | Description                                                                                                | Type                                                       | Default     |
-| ------------- | -------------- | ---------------------------------------------------------------------------------------------------------- | ---------------------------------------------------------- | ----------- |
-| `current`     | `current`      | ID of the current active item.                                                                             | `string`                                                   | `undefined` |
-| `data`        | `data`         | The data to build the menu (as an array of `MenuItem` or a JSON array) or the url to fetch to retrieve it. | `MenubarItem[] \| Promise<MenubarItem[]> \| URL \| string` | `undefined` |
-| `searchQuery` | `search-query` | Initial search query.                                                                                      | `string`                                                   | `undefined` |
+| Property      | Attribute      | Description                                                                                                   | Type                                                       | Default     |
+| ------------- | -------------- | ------------------------------------------------------------------------------------------------------------- | ---------------------------------------------------------- | ----------- |
+| `current`     | `current`      | ID of the current active item.                                                                                | `string`                                                   | `undefined` |
+| `data`        | `data`         | The data to build the menu (as an array of `MenubarItem` or a JSON array) or the url to fetch to retrieve it. | `MenubarItem[] \| Promise<MenubarItem[]> \| URL \| string` | `undefined` |
+| `searchQuery` | `search-query` | Initial search query.                                                                                         | `string`                                                   | `undefined` |
 
 
 ## Events
@@ -36,6 +36,7 @@ When a main menubar item is the current active one, a sub-menubar is shown and e
 graph TD;
   zanit-menubar --> zanit-mobile-menubar
   zanit-menubar --> z-icon
+  zanit-mobile-menubar --> z-logo
   zanit-mobile-menubar --> z-icon
   style zanit-menubar fill:#f9f,stroke:#333,stroke-width:4px
 ```
