+++ stencil.config.ts
@@ -23,4 +23,7 @@ export const config: Config = {
       serviceWorker: null, // disable service workers
     },
   ],
+  extras: {
+    enableImportInjection: true,
+  },
 };
